Fix artifact with concurrent add/remove LayoutTransitions
There was a bug with LayoutTransitions where if the animations
were of different duration (such as in the current status bar clock),
it was possible to end up in a bad situation because the previous animation
might outlast the new animation, causing the opposite end result
from what was expected. The fix was to keep track of the current
add/remove animations and to cancel any running ones prior to starting
new ones.

Change-Id: I884ce33ce0671f6ba6153ee3951c8d14c5ed5714