Error popup no longer flickers in TextViews.
The removed lines were committed by the Android Open Source Project.
Their intent was probably: the message was there before, it is identical
after a text change, let's remove it to not annoy the user who already saw it.

The behavior however is that the message is displayed then hidden, then displayed
as the user types.

Bug 3365016

Change-Id: Ie820f8e5465ad8ab5890272c42627686e0d7961b