Fix for TextView's error popup behavior when using soft keyboard.
Bug 3370191

The documented behavior is to hide the error when the text changes.
However, this should not be the case if the error was reset by a text watcher.

Comparing errorBefore and errorAfter as was done before is not sufficient in the
case where the error is reset to the same value. String pool optimization will re-use
the same Object and it will look like the error has not been modified (hence the
blinking behavior reported in the bug).

For this reason, TextView has a mErrorWasChanged flag. The fix is to export methods
that can use this flag as in done inside TextView when a physical keyboard is used.

These methods are hidden.

Change-Id: Ie3ec59a368f3b1588b81242890b971ac48e8ff7e