Fix keyboard hidden flag.
Configuration.hardKeyboardHidden should be set to
HARDKEYBOARDHIDDEN_YES when no keyboard is present or when it is
disabled.  The code in PhoneWindowManager.adjustConfigurationLw
does the right thing.  It also takes care of setting keyboardHidden
correctly when only a soft keyboard is available.

So it turns out that all we need to do to disable the hard keyboard
is to set Configuration.keyboard to KEYBOARD_NOKEYS *before* asking
the policy to adjust the configuration.

Bug: 3406101
Change-Id: I3519d02af30d8068198cb3d6b8fdde601c2fe8b4