The tab key completes text in AutoCompleteTextView.
Bug 3330651.

The first item in the list is not selected by default. From discussions
in other related bugs, there does not seem to be an agreement on this.

Supporting the actual token separator has also been punted. This would
require a new method in the Tokenizer, which could be ill-defined for
exotic tokenizer, plus typing a comma (for instance) to achieve a
completion is not a common pattern.

Change-Id: I30baf62077c412256175f871d21f4841e104f212