Fix bug 3413800 - Revise AutoCompleteTextView dropdown behavior
Be more conservative with when we let an AutoCompleteTextView's
dropdown box of completion suggestions cover the IME.

Disable the expand-when-touched behavior of the dropdown list when
more than 3 items can be seen at a time without it.

Don't let a ListPopupWindow that is expanding in response to touch
scroll the anchor view within its parent and slide the dropdown out
from under the user's finger.

Change-Id: I009accfd4e841c9a5e1072735d8a0b067a0bc06a