Fixes around cursor blink.
Bug 3394800

A previous fix called cancel when the window was detached. The cancel/uncancel
mechanism does not actually removes the Blink runnable.

It is indeed more a suspend, which is used when the window loses focus.
The problem here was that uncancel was never called.

Removing the runnable callback instead.

Also rationalized the use of makeBlink and the setting of mShowCursor

Change-Id: I92aac43a891991b7cc98738de0f12332ab16907a